requirements.txt
hydra-core==1.3.2
numpy==1.24.4
omegaconf==2.3.0
pandas==1.5.3
pyarrow==14.0.1
pytorch-lightning==1.6.4
scikit-learn==1.3.2
scipy==1.10.1
wandb==0.16.0
